### 2024-07-18 #7007 @editorialbot
[PRE REVIEW]: LorenzCycleToolkit: A Comprehensive Python Tool for Analyzing Atmospheric Energy Cycles
**Submitting author:** @daniloceano (<a href="http://orcid.org/0000-0003-4121-7583">Danilo Couto de Souza</a>)
**Repository:** https://github.com/daniloceano/LorenzCycleToolkit
**Branch with paper.md** (empty if default branch): joss-submission
**Version:** v1.0.0
