### 2024-07-18 #7007 @kthyng
@daniloceano Can you take a look to see what's breaking the paper compilation?
