### 2024-07-26 #7007 @kthyng
Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file.
