### 2024-07-27 2c46726
Fixed capitalization in the references
