+++ paper/paper.bib
@@ -1,19 +1,19 @@
 @article{lorenz1955,
     author = {Edward N. Lorenz},
-    title = {Available Potential Energy and the Maintenance of the General Circulation},
+    title = {{Available Potential Energy and the Maintenance of the General Circulation}},
     journal = {Tellus},
     volume = {7},
     number = {2},
     pages = {157-167},
     year  = {1955},
-    publisher = {Taylor & Francis},
+    publisher = {Taylor \& Francis},
     doi = {10.3402/tellusa.v7i2.8796},
     URL = {https://doi.org/10.3402/tellusa.v7i2.8796},
     eprint = {https://doi.org/10.3402/tellusa.v7i2.8796}
 }
 
 @article{ulbrich1991global,
-  title={The global energy cycle of stationary and transient atmospheric waves: results from ECMWF analyses},
+  title={{The Global Energy Cycle of Stationary and Transient Atmospheric Waves: Results from ECMWF Analyses}},
   author={Ulbrich, U and Speth, P},
   journal={Meteorology and Atmospheric Physics},
   volume={45},
@@ -23,7 +23,7 @@
 }
 
 @article{li2007lorenz,
-  title={Lorenz energy cycle of the global atmosphere based on reanalysis datasets},
+  title={{Lorenz Energy Cycle of the Global Atmosphere Based on Reanalysis Datasets}},
   author={Li, Liming and Ingersoll, Andrew P and Jiang, Xun and Feldman, Daniel and Yung, Yuk L},
   journal={Geophysical Research Letters},
   volume={34},
@@ -33,9 +33,9 @@
 }
 
 @article{boer2008energy,
-  title={The energy cycle in atmospheric models},
+  title={{The Energy Cycle in Atmospheric Models}},
   author={Boer, George J and Lambert, S},
-  journal={Climate dynamics},
+  journal={Climate Dynamics},
   volume={30},
   pages={371--390},
   year={2008},
@@ -43,15 +43,15 @@
 }
 
 @article{lorenz1967nature,
-  title={The nature and theory of the general circulation of the atmosphere},
+  title={{The Nature and Theory of the General Circulation of the Atmosphere}},
   author={Lorenz, Edward},
-  journal={World meteorological organization},
+  journal={World Meteorological Organization},
   volume={161},
   year={1967}
 }
 
 @article{veiga2013global,
-  title={A global and hemispherical analysis of the Lorenz energetics based on the representative concentration pathways used in CMIP5},
+  title={{A Global and Hemispherical Analysis of the Lorenz Energetics Based on the Representative Concentration Pathways Used in CMIP5}},
   author={Veiga, Jos{\'e} Augusto P and Ambrizzi, Tercio},
   journal={Advances in Meteorology},
   volume={2013},
@@ -62,7 +62,7 @@
 }
 
 @article{dias2011energy,
-  title={The energy cycle and structural evolution of cyclones over southeastern South America in three case studies},
+  title={{The Energy Cycle and Structural Evolution of Cyclones over Southeastern South America in Three Case Studies}},
   author={Dias Pinto, Joao Rafael and Da Rocha, Rosmeri Porf{\'\i}Rio},
   journal={Journal of Geophysical Research: Atmospheres},
   volume={116},
@@ -72,7 +72,7 @@
 }
 
 @article{michaelides1999quasi,
-  title={Quasi-Lagrangian energetics of an intense Mediterranean cyclone},
+  title={{Quasi-Lagrangian Energetics of an Intense Mediterranean Cyclone}},
   author={Michaelides, Silas C and Prezerakos, Nicholas G and Flocas, Helena A},
[…]
