### 2024-07-27 190e407
Add CircleCI configuration for build, test, and publish
