### 2024-07-27 043101c
Check if the build step correctly creates the wheel file
