+++ .circleci/config.yml
@@ -12,6 +12,10 @@ jobs:
           command: python3 setup.py sdist bdist_wheel
 
       - run:
+          name: Check if wheel file exists (build_test)
+          command: ls -l dist/
+
+      - run:
           name: Install dependencies (build_test)
           command: |
             python -m venv venv
@@ -50,6 +54,10 @@ jobs:
           command: python3 setup.py sdist bdist_wheel
 
       - run:
+          name: Check if wheel file exists (test_pypi_publish)
+          command: ls -l dist/
+
+      - run:
           name: Install dependencies (test_pypi_publish)
           command: |
             python -m venv venv
@@ -77,6 +85,10 @@ jobs:
       - run:
           name: Create whl (pypi_publish)
           command: python3 setup.py sdist bdist_wheel
+
+      - run:
+          name: Check if wheel file exists (pypi_publish)
+          command: ls -l dist/
 
       - run:
           name: Install dependencies (pypi_publish)
