+++ setup.py
@@ -13,7 +13,7 @@ def read_requirements():
 
 setup(
     name='LorenzCycleToolkit',
-    version='1.0.0',
+    version='1.0.1',
     packages=find_packages(),
     include_package_data=True,
     install_requires=read_requirements(),
