+++ CHANGELOG.md
@@ -1,0 +1,20 @@
+# Changelog
+
+All notable changes to this project will be documented in this file.
+
+## [1.0.1] - 2024-07-27
+### Added
+- Added `CHANGELOG.md` file to document changes.
+- Published project to PyPI.
+- Integrated CircleCI for continuous integration and deployment.
+
+## [1.0.0] - 2024-07-16
+### Initial Release
+- Initial release of LorenzCycleToolkit.
+- Implemented core functionalities for calculating the Lorenz Energy Cycle (LEC) in specific atmospheric regions.
+- Added support for both Eulerian and Semi-Lagrangian frameworks.
+- Provided options for fixed and moving computational domains.
+- Generated detailed results in CSV format for energy, conversion, and generation terms.
+- Included automatic plot generation for spatial and temporal analysis.
+- Utilized Xarray, Dask, Pandas, and MetPy for data handling and computations.
+- Comprehensive documentation and usage examples.
