+++ docs/source/installation.rst
@@ -4,26 +4,39 @@ Installation
 Prerequisites
 -------------
 - Python 3.12 or later
-- Required Python packages listed in `requirements.txt`
 
 Steps
 -----
+
+1. Install via pip::
+
+   pip install LorenzCycleToolkit
+
+2. Alternatively, clone the repository::
+
+   git clone https://github.com/daniloceano/lorenz-cycle.git
+   cd lorenz-cycle
+
+3. Create a virtual environment and activate it::
+
+   python -m venv venv
+   source venv/bin/activate
+
+4. Install the dependencies::
+
+   pip install -r requirements.txt
+
+For Conda users:
+
 1. Clone the repository::
 
    git clone https://github.com/daniloceano/lorenz-cycle.git
    cd lorenz-cycle
 
-2. Create and activate a virtual environment:
+2. Create a Conda environment and activate it::
 
-   - Using `venv`::
-
-      python -m venv venv
-      source venv/bin/activate
-
-   - Or using `conda`::
-
-      conda create --name lorenz-cycle python=3.12
-      conda activate lorenz-cycle
+   conda create --name lorenzcycletoolkit python=3.12
+   conda activate lorenzcycletoolkit
 
 3. Install the dependencies::
 
