### 2024-07-27 #7007 @daniloceano
@kthyng:

> Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file.

Thank you for your feedback. I have made an attempt to address the capitalization issue in the titles by enclosing key terms and acronyms in curly braces {}. However, I am not entirely sure if this fully resolves the concern. Could you please clarify if there are any specific instances or additional elements that need to be adjusted?

> @daniloceano also:
>
> * can you make a release on PyPI or is there a technical difficulty with doing so?

Regarding the PyPI release, it is now completed. You can find it at https://pypi.org/project/LorenzCycleToolkit/. Please note that I encountered some minor issues along the way, which necessitated updating the version to 1.0.3. Should I pull the changes from the main branch to the submission branch?
