### 2024-08-07 #7007 @kthyng
> Should I pull the changes from the main branch to the submission branch?

Since the review hasn't started, yes I think get everything into the review branch as up to date as possible so that it is all reviewed.

> Could you please clarify if there are any specific instances or additional elements that need to be adjusted?

I don't see anything that needs to be capitalized now — I generally check for proper nouns, etc.
