### 2024-08-07 #7007 @daniloceano
> Since the review hasn't started, yes I think get everything into the review branch as up to date as possible so that it is all reviewed.

I’ve pulled all the latest changes from the main branch into the joss-submission branch. Thank you.
