### 2024-08-26 #7139 @observingClouds
Hi @daniloceano,
This is the start of the review process. While the reviewers will be busy with reviewing your submission, could you please check the above mentioned references in your manuscript that do appear to miss a DOI and add it if possible. If no DOI is available for an entry, please add a URL and mention here why there is no DOI available, e.g. it is a reference to a poster.
Cheers,
Hauke
