+++ paper/paper.bib
@@ -19,7 +19,8 @@
   volume={45},
   pages={125--138},
   year={1991},
-  publisher={Springer}
+  publisher={Springer},
+  doi = {10.1007/bf01029650}
 }
 
 @article{li2007lorenz,
@@ -29,7 +30,8 @@
   volume={34},
   number={16},
   year={2007},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1029/2007GL029985}
 }
 
 @article{boer2008energy,
@@ -39,7 +41,8 @@
   volume={30},
   pages={371--390},
   year={2008},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s00382-007-0303-4}
 }
 
 @article{lorenz1967nature,
@@ -58,7 +61,8 @@
   number={1},
   pages={485047},
   year={2013},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1155/2013/485047}
 }
 
 @article{dias2011energy,
@@ -68,7 +72,8 @@
   volume={116},
   number={D14},
   year={2011},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1029/2011JD016217}
 }
 
 @article{michaelides1999quasi,
@@ -79,7 +84,8 @@
   number={553},
   pages={139--168},
   year={1999},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1002/qj.49712555309}
 }
 
 @article{rew1990netcdf,
@@ -90,7 +96,8 @@
   number={4},
   pages={76--82},
   year={1990},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/38.56302}
 }
 
 @article{may2022metpy,
@@ -100,7 +107,8 @@
   volume={103},
   number={10},
   pages={E2273--E2284},
-  year={2022}
+  year={2022},
+  doi={10.1175/BAMS-D-21-0125.1}
 }
 
 @book{daniel2019data,
@@ -117,14 +125,16 @@
   volume={5},
   number={1},
   pages={10--10},
-  year={2017}
+  year={2017},
+  doi={10.5334/jors.148}
 }
 
 @article{reback2020pandas,
   title={{pandas-dev/pandas: Pandas 1.0.5}},
   author={Reback, Jeff and McKinney, Wes and Van Den Bossche, Joris and Augspurger, Tom and Cloud, Phillip and Klein, Adam and Hawkins, Simon and Roeschke, Matthew and Tratner, Jeff and She, Chang and others},
   journal={Zenodo},
-  year={2020}
+  year={2020},
+  doi={10.5281/zenodo.3898987 }
 }
 
 @article{harris2020array,
@@ -135,5 +145,6 @@
   number={7825},
   pages={357--362},
   year={2020},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
+  doi={10.1038/s41586-020-2649-2}
 }
