### 2024-08-26 #7139 @daniloceano
> Hi @daniloceano, This is the start of the review process. While the reviewers will be busy with reviewing your submission, could you please check the above mentioned references in your manuscript that do appear to miss a DOI and add it if possible. If no DOI is available for an entry, please add a URL and mention here why there is no DOI available, e.g. it is a reference to a poster. Cheers, Hauke

Hi Hauke,

Thank you for initiating the review process.

I have provided the DOIs for all references except for the following two, as they are books:

- "The Nature and Theory of the General Circulation of the Atmosphere"
- "Data Science with Python and Dask"

Since these are books, they do not have associated DOIs.

Please let me know if further adjustments are needed.

Best regards,
Danilo
