### 2024-08-26 #7139 @observingClouds
@daniloceano thanks for the update. Could you please add the ISBN for those books? We just like to have some permanent identifiers referenced to make it easier for readers to find the references later.
