### 2024-08-26 #7139 @amylu00
Hi @daniloceano! This seems like a cool repo, thank you for working on it. A few small comments/questions as I'm going through the review:
1.  What is the best way to run the test files? It seems like when I try running any of the test files, I get an error saying `ModuleNotFoundError: No module named 'lorenzcycletoolkit`.
2. The second example ("Moving (Semi-Lagrangian) Framework Example") in your documentation required creating a file named _track_ in the inputs folder. It would be great if that was already created within the repo and/or edit the documentation to say _inputs/track_ file instead of _inputs/track_file_. This way, it is consistent with the first example where the user does not have to create the file themselves and also with respect to the naming of the files.
3. The guidelines for any third party that would like to contribute to the software is clear. I would suggest including guidelines for those seeking support and also how and where to report any issues with the software.
