+++ paper/paper.bib
@@ -50,7 +50,8 @@
   author={Lorenz, Edward},
   journal={World Meteorological Organization},
   volume={161},
-  year={1967}
+  year={1967},
+  isbn={9788873954736},
 }
 
 @article{veiga2013global,
@@ -115,7 +116,8 @@
   title={{Data Science with Python and Dask}},
   author={Daniel, Jesse},
   year={2019},
-  publisher={Simon and Schuster}
+  publisher={Simon and Schuster},
+  isbn={9781638353546}
 }
 
 @article{hoyer2017xarray,
