### 2024-08-28 #7139 @daniloceano
> @daniloceano thanks for the update. Could you please add the ISBN for those books? We just like to have some permanent identifiers referenced to make it easier for readers to find the references later.

Thank you for your feedback. I've added the ISBN for "Data Science with Python and Dask." However, "The Nature and Theory of the General Circulation of the Atmosphere" was published before ISBN numbers were introduced, so it does not have one. I was only able to find an ISBN for an Italian version of the book.
