### 2024-08-28 #7139 @daniloceano
> Hi @daniloceano! This seems like a cool repo, thank you for working on it. A few small comments/questions as I'm going through the review:
>
> 1. What is the best way to run the test files? It seems like when I try running any of the test files, I get an error saying `ModuleNotFoundError: No module named 'lorenzcycletoolkit`.
> 2. The second example ("Moving (Semi-Lagrangian) Framework Example") in your documentation required creating a file named _track_ in the inputs folder. It would be great if that was already created within the repo and/or edit the documentation to say _inputs/track_ file instead of _inputs/track_file_. This way, it is consistent with the first example where the user does not have to create the file themselves and also with respect to the naming of the files.
> 3. The guidelines for any third party that would like to contribute to the software is clear. I would suggest including guidelines for those seeking support and also how and where to report any issues with the software.

Hi @amylu00,

Thank you so much for taking the time to review the repository! I appreciate your feedback and am happy to address your comments.

1. ModuleNotFoundError: How are you currently running the test files? The updated documentation now indicates that the program should be run from the top-level directory of the project. This ensures that the modules are correctly found and executed.
2. Track File Creation: Great point! There is already a sample track file included in the repository, and it can be easily copied to inputs/track. I’ve updated the documentation to reflect this, so users will know to use the sample file instead of creating it from scratch.
3. Support and Issue Reporting: I’ve also added a section to the documentation outlining the best ways to seek support and report any issues with the software. This should make it easier for users to know where and how to get help.

Please let me know if you have any further questions or if there’s anything else you’d like me to clarify!
