### 2024-08-28 #7139 @amylu00
> 1. ModuleNotFoundError: How are you currently running the test files? The updated documentation now indicates that the program should be run from the top-level directory of the project. This ensures that the modules are correctly found and executed.

I am running from the top-level directory and used `python tests/test_ERA5_track.py`. I tried running `python lorenzcycletoolkit.py tests/test_ERA5_track.py` as well but then I think that runs the `lorenzcycletoolkit.py` module instead of the `tests/test_ERA5_track.py` one.

Documentation looks good to me now, thanks for adding clarification!
