### 2024-08-28 #7139 @daniloceano
> I am running from the top-level directory and used `python tests/test_ERA5_track.py`. I tried running `python lorenzcycletoolkit.py tests/test_ERA5_track.py` as well but then I think that runs the `lorenzcycletoolkit.py` module instead of the `tests/test_ERA5_track.py` one.

Thanks for the update! Just to clarify so I can replicate the issue on my end—did you install the package using pip, or have you only cloned the repository and are running it directly from there?
