### 2024-08-29 #7139 @daniloceano
Hi @amylu00,

Could you please run the following tests for me?

1. From the top-level directory, run `pytest`.
2. From the top-level directory, run python `lorenzcycletoolkit.py`.

If both commands work without issues, it seems everything is functioning correctly. I can then update the documentation to indicate that this is the proper way to run the tests.

Thanks!
