### 2024-09-04 #7139 @einaraz
Hi @daniloceano! Great job on the project and the software! I successfully installed and ran all the examples. I have a few comments and suggestions:

1. The examples in the tutorial were clear. However, it might be helpful to move (or repeat) the definition of the Flags in the Tutorial and/or Usage section to make it more linear and clear for users when running the examples.
2. The "contributing guide" page is currently empty.
3. The README mentions that the license is MIT, but it appears that the actual license is GNU.
4. I wonder if the README file should contain installation instructions. At least clarify that installation instructions are included in the documentation.
5. Since the paper itself has no space for results, I think it would be great somewhere in the repo (README or in the results section of the documentation) to describe/interpret some of the outputs. Ofc that's only a suggestion, but I think it would make the project more appealing and help users understand the figures.
