+++ docs/source/license.rst
@@ -1,4 +1,4 @@
 License
 =======
 
-This project is licensed under the GNU License Version 3 - see the `LICENSE <LICENSE>`_ file for details.
+This project is licensed under the GNU License Version 3 - see the `LICENSE <../LICENSE>`_ file for details.
