+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.0.5] - 2024-09-04
+
+### Added
+- Updated README to clarify that installation instructions are included in the documentation.
+- Revised README and documentation to reflect the GNU license, correcting the previously stated MIT license.
+- Moved the "Flags" section to the "Usage" section to improve tutorial clarity and make the process more linear for users.
+- Added a note in the documentation explaining that results generated using the LorenzCycleToolkit will be referenced in a future journal publication to aid users in interpreting outputs.
+
+### Fixed
+- Corrected links to the Contributing Guide and License in the README and documentation.
+
 ## [1.0.4] - 2024-08-28
 
 ### Added
