### 2024-09-05 #7139 @daniloceano
Hi @einaraz. Thank you very much for the feedback.

> 1\. The examples in the tutorial were clear. However, it might be helpful to move (or repeat) the definition of the Flags in the Tutorial and/or Usage section to make it more linear and clear for users when running the examples.

I've moved the Flags section into the "Usage" section to make the flow more straightforward for users. Does this look better to you?

> 2\. The "contributing guide" page is currently empty.

Could you please check again? The guide should now be accessible here: https://daniloceano.github.io/LorenzCycleToolkit/contributing.html

> 3\. The README mentions that the license is MIT, but it appears that the actual license is GNU.

This has been corrected. Thanks for catching that!

> 4\. I wonder if the README file should contain installation instructions. At least clarify that installation instructions are included in the documentation.

I’ve updated the README to mention that installation instructions can be found in the documentation.

> 5\. Since the paper itself has no space for results, I think it would be great somewhere in the repo (README or in the results section of the documentation) to describe/interpret some of the outputs. Ofc that's only a suggestion, but I think it would make the project more appealing and help users understand the figures.

You’re right — interpreting Lorenz Energy Cycle outputs requires a solid understanding of atmospheric dynamics, and a full explanation would typically be found in textbooks. However, I’m currently preparing a submission of results generated using this program for a dedicated journal. Once the paper is published, I’ll link it in the results section of the documentation to provide further context and help future users interpret the outputs of LorenzCycleToolkit.

In the meantime, I encourage users to refer to the original Lorenz (1965) paper and the references in the documentation to better understand the energy cycle and its applications.
