### 2024-09-05 #7139 @einaraz
Thank you for addressing my comments, @daniloceano! The proposed changes look good!

1. I see you updated the doc files to add Flags under usage; however, the documentation was not updated. Could you please rebuild the page?
2. Also, the link for contributing you provided in the reply directs to the correct place; the one linked on the README file, under Contributing, is trying to link directly to the file 'CONTRIBUTING.md' that does not exist in the directory.
3. I think it's a great idea to direct users to your other publications and additional papers on the subject! A suggestions would be to add a "References" session at the bottom of the README file to include relevant materials.
