### 2024-09-05 3f35f9b
Version 1.0.6
