### 2024-09-05 #7139 @daniloceano
Thank you for your feedback and suggestions!

1) I have rebuilt the page to reflect the updated documentation with the changes to the "Flags" section under "Usage."

2) I’ve corrected the link in the README.md. The link now points to the correct location in the documentation for contributing guidelines.

3) Regarding the references, thank you for the suggestion! I've included all relevant references and materials within the main documentation instead of the README.md. This keeps the README concise, while the documentation provides a more in-depth reference list.

Please let me know if further changes are needed!
