+++ paper/paper.md
@@ -31,13 +31,13 @@ bibliography: paper.bib
 
 # Summary
 
-The LorenzCycleToolkit is a Python package designed for the calculation and analysis of the Lorenz Energy Cycle (LEC) in specified atmospheric regions. Originally proposed by Edward Lorenz in 1965, the LEC provides a comprehensive framework for understanding the transformation and distribution of energy within the atmosphere, particularly focusing on the zonal and eddy components of Kinetic and Available Potential Energy. This toolkit enables researchers to study these energy transformations in various atmospheric phenomena, thereby enhancing our understanding of atmospheric dynamics and improving weather and climate prediction models.
+The `LorenzCycleToolkit` is a Python package designed for the calculation and analysis of the Lorenz Energy Cycle (LEC) in specified atmospheric regions. Originally proposed by Edward Lorenz in 1965, the LEC provides a comprehensive framework for understanding the transformation and distribution of energy within the atmosphere, particularly focusing on the zonal and eddy components of kinetic and available potential energy. This toolkit enables researchers to study these energy transformations in various atmospheric phenomena, thereby enhancing our understanding of atmospheric dynamics and improving weather and climate prediction models.
 
-The LorenzCycleToolkit supports both Eulerian and Semi-Lagrangian frameworks, allowing for flexibility in the analysis of atmospheric data. It can handle reanalysis or model data provided in isobaric levels, and outputs results in CSV format for easy analysis. Additionally, the toolkit offers options for automatic plot generation, including spatial representations of the computational domain, time series, Hovmöller diagrams, and complete energy flux boxes for the LEC. This package is a valuable tool for meteorologists and climate scientists aiming to diagnose and improve the performance of numerical weather prediction models and to study the impacts of climate change on atmospheric energy processes.
+The `LorenzCycleToolkit` supports both Eulerian and semi-Lagrangian frameworks, allowing for flexibility in the analysis of atmospheric data. It can handle reanalysis or model data provided on isobaric levels, and outputs results in the CSV format for easy analysis. Additionally, the toolkit offers options for automatic plot generation, including spatial representations of the computational domain, time series, Hovmöller diagrams, and complete energy flux boxes for the LEC. This package is a valuable tool for meteorologists and climate scientists aiming to diagnose and improve the performance of numerical weather prediction models and to study the impacts of climate change on atmospheric energy processes.
 
 # Background
 
-The Lorenz Energy Cycle (LEC), introduced by [@lorenz1955], is a framework to describe the processes of energy conversion and transfer within the Earth's atmosphere. It serves as a diagnostic tool for the general circulation of the atmosphere, providing insights into its dynamic and thermodynamic structures, and information about atmospheric flows and the transport of heat, moisture, and angular momentum. The LEC divides atmospheric energy into four main reservoirs: available potential energy (APE), kinetic energy (K), and their respective zonal mean (Z) and eddy (E) components. The LEC can indicate how energy is generated, converted, and dissipated within these reservoirs.
+The Lorenz Energy Cycle (LEC), introduced by @lorenz1955, is a framework to describe the processes of energy conversion and transfer within the Earth's atmosphere. It serves as a diagnostic tool for the general circulation of the atmosphere, providing insights into its dynamic and thermodynamic structures, and information about atmospheric flows and the transport of heat, moisture, and angular momentum. The LEC divides atmospheric energy into four main reservoirs: available potential energy (APE), kinetic energy (K), and their respective zonal mean (Z) and eddy (E) components. The LEC can indicate how energy is generated, converted, and dissipated within these reservoirs.
 
[…]
