### 2024-09-15 #7139 @observingClouds
Hi @daniloceano,
There are now a few last items that we need to do as mentioned in the checklist below. Please also have a look at my proposed changes to the manuscript.
Cheers!
