### 2024-09-15 #7139 @daniloceano
@observingClouds

Everything's done!

I have double-checked the authors and affiliations, created the v1.0.7 release on GitHub, and archived the release on Zenodo. The authors list, ORCIDs, and license in the Zenodo repository match those in the JOSS paper.

The Zenodo DOI is: [10.5281/zenodo.13765959](https://doi.org/10.5281/zenodo.13765959)
