### 2024-09-24 #7139 @kthyng
Everything looks good except one small thing @daniloceano: can you remove the extra "figure 1", "figure 2" from your captions?
