### 2024-09-25 c16e1da
Removed captions from figures
