+++ paper/paper.md
@@ -39,7 +39,7 @@ The `LorenzCycleToolkit` supports both Eulerian and semi-Lagrangian frameworks, 
 
 The Lorenz Energy Cycle (LEC), introduced by @lorenz1955, is a framework to describe the processes of energy conversion and transfer within the Earth's atmosphere. It serves as a diagnostic tool for the general circulation of the atmosphere, providing insights into its dynamic and thermodynamic structures, and information about atmospheric flows and the transport of heat, moisture, and angular momentum. The LEC divides atmospheric energy into four main reservoirs: available potential energy (APE), kinetic energy (K), and their respective zonal mean (Z) and eddy (E) components. The LEC can indicate how energy is generated, converted, and dissipated within these reservoirs.
 
-![Figure 1: Schematic representation of the Lorenz Energy Cycle. The figure shows the four main energy reservoirs in the atmosphere: zonal available potential energy ($A_Z$), eddy available potential energy ($A_E$), zonal kinetic energy ($K_Z$), and eddy kinetic energy ($K_E$). The arrows indicate the energy transformations between these reservoirs: $C_Z$ (conversion from $A_Z$ to $K_Z$), $C_E$ (conversion from $A_E$ to $K_E$), $C_A$ (conversion between $A_Z$ and $A_E$), and $C_K$ (conversion between $K_Z$ and $K_E$). Additional terms include $B$ (boundary fluxes), $R$ (residuals), $G$ (generation terms), and $D$ (dissipation terms).](figures/LEC_example.png)
+![Schematic representation of the Lorenz Energy Cycle. The figure shows the four main energy reservoirs in the atmosphere: zonal available potential energy ($A_Z$), eddy available potential energy ($A_E$), zonal kinetic energy ($K_Z$), and eddy kinetic energy ($K_E$). The arrows indicate the energy transformations between these reservoirs: $C_Z$ (conversion from $A_Z$ to $K_Z$), $C_E$ (conversion from $A_E$ to $K_E$), $C_A$ (conversion between $A_Z$ and $A_E$), and $C_K$ (conversion between $K_Z$ and $K_E$). Additional terms include $B$ (boundary fluxes), $R$ (residuals), $G$ (generation terms), and $D$ (dissipation terms).](figures/LEC_example.png)
 
 Applications of the LEC are diverse and significant in meteorology and climate science. It aids in diagnosing the performance of numerical weather prediction models by comparing simulated and observed energy cycles [@ulbrich1991global]. Moreover, it is used to study climate change impacts, such as shifts in energy transfer processes due to global warming [@li2007lorenz; @veiga2013global]. Understanding the LEC is also crucial for improving the representation of atmospheric processes in general circulation models (GCMs), which are essential for accurate climate projections [@lorenz1967nature; @boer2008energy].
 
@@ -47,7 +47,7 @@ Applications of the LEC are diverse and significant in meteorology and climate s
 
 `LorenzCycleToolkit` is a Python package designed for calculating the Lorenz Energy Cycle (LEC) in limited atmospheric regions, supporting both Eulerian and semi-Lagrangian frameworks. The formulations used are the same as in @michaelides1999quasi. The Eulerian framework computes the LEC for a fixed region in the atmosphere [@dias2011energy], requiring users to provide a "box limits" file containing the minimum and maximum latitudes and longitudes for the desired computational domain. The semi-Lagrangian framework, on the other hand, uses a domain that changes position with time [@michaelides1999quasi], which is especially useful for studying cyclonic systems. For this option, users must specify the central position of the domain for each timestep in a "track" CSV file, which can also specify the domain width and length. If these specifications are not provided, a default 15°x15° domain is used. The `LorenzCycleToolkit` also offers an option for users to interactively choose the domain position and size for each data timestep.
 
[…]
