### 2024-09-25 #7139 @kthyng
(The DOI isn't resolving for me so I am leaving this issue open until I can verify it works — let me know if it works for any of you!)
