### 2024-09-25 #7139 @daniloceano
Thank you @kthyng!

The DOI is not resolving for me as well.
